Better octave representation in the UI
Former-commit-id: 302ace8914ea1ba822aec311989d8ee9ee39fd20

+++ src/Materials/Node/Blocks/cloudBlock.ts
@@ -11,7 +11,7 @@ import { Scene } from '../../../scene';
  */
 export class CloudBlock extends NodeMaterialBlock {
     /** Gets or sets the number of octaves */
-    @editableInPropertyPage("Octaves", PropertyTypeForEdition.Float)
+    @editableInPropertyPage("Octaves", PropertyTypeForEdition.Int)
     public octaves = 6.0;
 
     /**
